--- src/components/ui/Button.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Button = () => {
-  return <div></div>;
-};
-
-export default Button;
