--- src/components/ui/Card2.module.scss
@@ -1,49 +1,0 @@
-.container {
-  width: 100%;
-  max-width: 350px;
-  border-radius: 10px;
-  overflow: hidden;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  background: white;
-  display: flex;
-  flex-direction: column;
-  cursor: pointer;
-
-  > .image {
-    height: 320px;
-    background-size: cover;
-    background-position: center;
-  }
-
-  > .content {
-    padding: 16px;
-    display: flex;
-    flex-direction: column;
-
-    .title {
-      margin: 0;
-      color: #333;
-      font-size: 24px;
-      text-align: left;
-    }
-    .description {
-      margin: 10px 0;
-      color: #666;
-      font-size: 16px;
-      text-align: left;
-    }
-    .link {
-      color: black;
-      text-decoration: none;
-      font-size: 16px;
-      font-weight: normal;
-      text-align: left;
-      margin-top: 20px;
-    }
-  }
-  &:hover {
-    .link {
-      color: gray;
-    }
-  }
-}
