+++ src/pages/impact/sustainability/Double.module.scss
@@ -1,12 +1,7 @@
 .container {
-  position: relative;
-  margin-top: 120px;
-  display: block;
-  justify-content: space-between;
-  text-align: -webkit-center;
-  //   여기
   position: relative; /* 상대 위치 설정 */
-  overflow: hidden;
+  margin-top: 120px;
+  //   여기
 
   > .title {
     font-size: 32px;
@@ -17,11 +12,12 @@
   }
 
   > .campaign {
-    max-width: 1200px;
+    margin: 120px auto 0;
+    max-width: 1120px;
     display: flex;
     align-items: flex-start;
     justify-content: space-between;
-    margin-inline: -16px;
+    // margin-inline: -16px;
     padding: 0;
     // 여기
     position: relative;
@@ -32,9 +28,9 @@
       display: flex;
       flex-direction: column;
       justify-content: center;
-      border-radius: 8px;
+      border-radius: 16px;
       background-color: white;
-      box-shadow: 0 4px 8px rgba(0, 0, 0, 1);
+      box-shadow: 0 14px 18px rgba(0, 0, 0, 0.06);
       //   여기
       max-width: 600px;
       position: relative;
@@ -46,33 +42,43 @@
         line-height: 32px;
       }
       .contentDescription {
+        margin-top: 16px;
         font-size: 16px;
         line-height: 26px;
-        margin-bottom: 250px;
       }
       .contentLink {
-        max-width: 400px;
+        margin-top: 250px;
+        max-width: 327px;
         display: inline-block;
         padding: 8px 16px;
         border: 2px solid black;
         border-radius: 24px;
         text-decoration: none;
+        text-align: center;
         color: black;
         font-weight: bold;
         background-color: transparent;
+
+        &:hover {
+          border-color: #6e6e6e;
+          color: #6e6e6e;
+        }
       }
     }
 
     > .imageContainer {
+      margin-top: 80px;
       flex: 1;
-      padding: 64px 16px;
+      height: 640px;
+      height: 512px;
       display: flex;
       align-items: center;
       justify-content: center;
 
       .campaignImage {
-        width: 550px;
-        height: 550px;
+        border-radius: 16px;
+        width: 100%;
+        height: 100%;
         object-fit: cover;
       }
     }
