--- src/pages/impact/sustainability/tpgus/GlobalReport.module.scss
@@ -1,69 +1,0 @@
-.container {
-  > .title {
-    margin-top: 120px;
-    text-align: center;
-    font-size: 32px;
-    font-weight: 400;
-  }
-  > .section {
-    position: relative;
-    margin-top: 120px;
-    height: 592px;
-
-    > .info {
-      position: absolute;
-      z-index: 2;
-      display: flex;
-      flex-direction: column;
-      width: 50%;
-      background-color: white;
-      min-height: 552px;
-      padding: 64px 80px;
-      border-radius: 16px;
-
-      > .detail {
-        flex: 1;
-
-        h3 {
-          font-size: 24px;
-          line-height: 32px;
-        }
-        p {
-          margin-top: 16px;
-          font-size: 16px;
-          line-height: 26px;
-        }
-      }
-      button {
-        width: 327px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        color: #000;
-        border: 2px solid black;
-        background-color: transparent;
-        border-radius: 48px;
-        font-size: 14px;
-        font-weight: 600;
-        height: 40px;
-        line-height: 14px;
-        cursor: pointer;
-      }
-    }
-    > .imageWrapper {
-      margin-top: 80px;
-      width: 640px;
-      height: 512px;
-      position: absolute;
-      z-index: 1;
-      right: 0;
-
-      > img {
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-        border-radius: 16px;
-      }
-    }
-  }
-}
