--- src/pages/impact/sustainability/tpgus/index.module.scss
@@ -1,170 +1,0 @@
-@import '../../../../styles/common.scss';
-
-.container {
-  // Banner
-  > .banner {
-    width: 100%;
-    height: 640px;
-    position: relative;
-
-    > img {
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-    > .title {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      font-size: 40px;
-      color: white;
-      letter-spacing: 2px;
-    }
-  }
-  // Effort
-  > .effort {
-    display: flex;
-    flex-direction: column;
-    margin-top: 120px;
-
-    > .header {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-
-      > .title {
-        font-size: 38px;
-        font-weight: 500;
-      }
-      > .description {
-        margin-top: 24px;
-        font-weight: 500;
-        text-align: center;
-        line-height: 28px;
-      }
-    }
-    > .box {
-      display: flex;
-      justify-content: center;
-      margin-top: 24px;
-
-      > .content {
-        width: 1280px;
-        height: 592px;
-        position: relative;
-        text-align: center;
-
-        > img {
-          width: 100%;
-          height: 100%;
-          object-fit: cover;
-          border-radius: 16px;
-        }
-        > .responsibility {
-          position: absolute;
-          top: 0;
-          bottom: 0;
-          left: 0;
-          right: 0;
-          z-index: 1;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: flex-start;
-          background: linear-gradient(to right, rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.1));
-          border-radius: 16px;
-
-          > .text {
-            text-align: start;
-            margin-left: 14px;
-            font-weight: 500;
-            font-size: 16px;
-            color: white;
-            margin-left: 64px;
-          }
-          > a {
-            margin: 32px 0 0 64px;
-            width: 280px;
-            height: 40px;
-            background-color: white;
-            border-radius: 16px;
-            font-size: 14px;
-            font-weight: 500;
-            color: black;
-            text-decoration: none;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            &:hover {
-              background-color: rgba(255, 255, 255, 0.9);
-            }
-          }
-        }
-      }
-    }
-  }
-  // Fields
-  > .fields {
-    margin-top: 120px;
-
-    > .title {
-      text-align: center;
-      font-size: 32px;
-      font-weight: 400;
-    }
-    > .cards {
-      display: grid;
-      grid-template-columns: repeat(3, 1fr);
-      grid-gap: 32px;
-      margin: 40px 160px 0 160px;
-
-      > .card {
-        flex: 1;
-        border-radius: 16px;
-        overflow: hidden;
-        text-decoration: none;
-        color: black;
-        background-color: white;
-        transition: transform 0.3s;
-
-        &.highlight {
-          transform: scale(1.02);
-        }
-
-        > img {
-          width: 100%;
-        }
-        > .content {
-          padding: 40px;
-          font-weight: 500;
-
-          > .title {
-            font-size: 24px;
-            line-height: 32px;
-          }
-          > .description {
-            margin-top: 16px;
-            line-height: 26px;
-            font-size: 16px;
-            min-height: 120px;
-            color: #444444;
-          }
-          > .action {
-            display: flex;
-            align-items: center;
-            height: 40px;
-
-            &.highlight {
-              color: #666666;
-            }
-          }
-        }
-      }
-    }
-  }
-  > .globalReport {
-    margin: 40px 160px;
-  }
-}
